Split `model.rs` up into several files
No changes to the public API, we still re-export everything publicly at
its original path.
